Fixed pointer issues, general code cleanup
I LOVE POINTERS :)

+++ src/pvegas/pros_adapters/ProsADIPotentiometer.cpp
@@ -6,9 +6,7 @@ ProsADIPotentiometer::ProsADIPotentiometer(
     std::unique_ptr<pros::adi::AnalogIn>& potentiometer)
     : m_potentiometer{std::move(potentiometer)} {}
 
-void ProsADIPotentiometer::init() {
-  calibrate();
-}
+void ProsADIPotentiometer::init() { calibrate(); }
 
 void ProsADIPotentiometer::calibrate() {
   m_potentiometer->calibrate();
@@ -16,7 +14,7 @@ void ProsADIPotentiometer::calibrate() {
 }
 
 double ProsADIPotentiometer::getAngle() {
-  double decidegrees{m_potentiometer->get_value()};
+  int32_t decidegrees{m_potentiometer->get_value()};
   return decidegrees * DECIDEGREES_TO_RADIANS;
 }
 }  // namespace pros_adapters
